server/requirements.txt
Adafruit-Blinka==8.53.0
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-mpu6050==1.3.2
adafruit-circuitpython-register==1.10.2
adafruit-circuitpython-requests==4.1.9
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
AHRS==0.3.1
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
asttokens==3.0.0
attrs==25.1.0
binho-host-adapter==0.1.6
blinker==1.9.0
-e git+https://github.com/BracketBotCapstone/quickstart.git@6d24a57bf413931fd7a4dea1575b8d9ebaf4d758#egg=bracket_bot
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.4
contourpy==1.3.1
control==0.10.1
csaps==1.2.0
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
executing==2.2.0
fastask==0.4.7
filelock==3.17.0
Flask==3.1.0
fonttools==4.56.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
intelhex==2.3.0
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.8
lgpio==0.2.2.0
libtmux==0.42.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
monotonic==1.6
mpmath==1.3.0
navlie==0.0.1
networkx==3.4.2
numpy==1.26.4
odrive==0.5.1.post0
openai==1.61.1
opencv-python==4.11.0.86
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pyftdi==0.56.0
Pygments==2.19.1
pymlg @ git+https://github.com/decargroup/pymlg@e0bfac968180713aaf06a19a0ddb634a83945258
pyngrok==7.2.3
pyparsing==3.2.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyusb==1.3.1
pywavemap @ git+https://github.com/ethz-asl/wavemap.git@db4f2fffdc1da1bec042d51bfff58029c0c9c827#subdirectory=library/python
PyYAML==6.0.2
requests==2.32.3
rerun-sdk==0.22.0
rpi-lgpio==0.6
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
rpi_hardware_pwm==0.2.2
ruff==0.9.5
scikit-build==0.18.1
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
smbus2==0.5.0
sniffio==1.3.1
sshkeyboard==2.3.1
stack-data==0.6.3
sympy==1.13.1
sysv-ipc==1.1.0
toml==0.10.2
tomlkit==0.13.2
torch==2.6.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
